Split connections module to protocol (#670)
